+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [0.0.3] - 2018-08-25
+### Added
+- File `grammar.rb` with minimalist grammar.
+- Initial `Parser` class commit
+
+### Changed
+- Class`Tokenizer` recognizes `define` keyword
+- Spec file `Tokenizer_spec.rb` expanded with more tests.
+
 ## [0.0.2] - 2018-08-25
 ### Changed
 - Class`Tokenizer` improved, does recognize delimiters, booleans, integers, real numbers, strings, and identifiers.